requirements.txt
# MkDocs and theme
mkdocs>=1.5.0,<2.0.0
mkdocs-material>=9.4.0,<10.0.0

# Plugins
mkdocs-blog-plugin>=0.25,<1.0.0

# Image processing
pillow>=10.0.0,<11.0.0
cairosvg>=2.7.0,<3.0.0

# YAML processing
pyyaml>=6.0,<7.0.0

# Validation and testing tools
# Note: Using npm-based tools for link checking instead of Python packages